Release 2.0.0

+++ Raphael.Net/SolverInput.cs
@@ -2,8 +2,8 @@ namespace Raphael;
 
 public readonly record struct SolverInput
 {
-    public short CP { get; init; }
-    public sbyte Durability { get; init; }
+    public ushort CP { get; init; }
+    public ushort Durability { get; init; }
     public ushort Progress { get; init; }
     public ushort Quality { get; init; }
     public ushort BaseProgressGain { get; init; }
